add Point class; release/debug option in cmake

+++ include/Tools.hpp
@@ -6,8 +6,8 @@
 using namespace std;
 
 // Product of all (numeric) elements within the vector
-template <typename T_num> T_num elemProd (vector<T_num>& v, unsigned int arrFrom, unsigned int arrTo);
-template <typename T_num> T_num elemProd (vector<T_num>& v, unsigned int arrFrom);
+template <typename T_num> T_num elemProd (vector<T_num>& v, size_t arrFrom, size_t arrTo);
+template <typename T_num> T_num elemProd (vector<T_num>& v, size_t arrFrom);
 template <typename T_num> T_num elemProd (vector<T_num>& v);
