+++ src/Tools.cpp
@@ -11,7 +11,7 @@ using namespace std;
 
 // Product of all (numeric) elements within the vector
 template <typename T_num>
-T_num elemProd (vector<T_num>& v, unsigned int arrFrom, unsigned int arrTo) {
+T_num elemProd (vector<T_num>& v, size_t arrFrom, size_t arrTo) {
 	T_num ret;
 
 	if (v.size() <= arrFrom) {
@@ -20,21 +20,21 @@ T_num elemProd (vector<T_num>& v, unsigned int arrFrom, unsigned int arrTo) {
 
 	ret = v[arrFrom];
 
-	for (int i = arrFrom+1; i < arrTo; i++) {
+	for (size_t i = arrFrom+1; i < arrTo; i++) {
 		ret *= v[i];
 	}
 
 	return ret;
 }
 template <typename T_num>
-T_num elemProd (vector<T_num>& v, unsigned int arrFrom) {
-	return elemProd (v, arrFrom, (unsigned int)(v.size()));
+T_num elemProd (vector<T_num>& v, size_t arrFrom) {
+	return elemProd (v, arrFrom, (size_t)(v.size()));
 }
 template <typename T_num>
 T_num elemProd (vector<T_num>& v) {
-	return elemProd (v, (unsigned int)0, (unsigned int)(v.size()));
+	return elemProd (v, (size_t)0, (size_t)(v.size()));
 }
 
-template int elemProd<int> (vector<int>& v, unsigned int arrFrom, unsigned int arrTo);
-template int elemProd<int> (vector<int>& v, unsigned int arrFrom);
+template int elemProd<int> (vector<int>& v, size_t arrFrom, size_t arrTo);
+template int elemProd<int> (vector<int>& v, size_t arrFrom);
 template int elemProd<int> (vector<int>& v);
